RegistryActivity.jl Documentation

This is the reference documentation of RegistryActivity.jl.

Index

RegistryActivity._non_yanked_versions

RegistryActivity.checkout_date

RegistryActivity.clone_registry

RegistryActivity.count_packages

RegistryActivity.count_versions

RegistryActivity.extrema_dates

RegistryActivity.general_registry

RegistryActivity.get_date

RegistryActivity.months_list

RegistryActivity.registry_activity

Functions

RegistryActivity._non_yanked_versions — Method

_non_yanked_versions(file)

Get the list of non-yanked versions of the package for which file is the path to its Versions.toml.

RegistryActivity.checkout_date — Method

checkout_date(path, date::Date; branch="master")

Check out branch of Git repository at path on date.

RegistryActivity.clone_registry — Method

clone_registry(; origin="https://github.com/JuliaRegistries/General.git", dest=mktempdir())

Do a git clone of origin to dest.

RegistryActivity.count_packages — Function

count_packages(path, packages=TOML.parsefile(joinpath(path, "Registry.toml"))["packages"])

Count the number of packages in the registry at path, using the list of packages.

RegistryActivity.count_versions — Function

count_versions(path, packages=TOML.parsefile(joinpath(path, "Registry.toml"))["packages"])

Count the number of versions in the registry at path, using the list of packages.

RegistryActivity.extrema_dates — Method

extrema_dates(path)

Get the first and last date of commits in the git repository at path.

RegistryActivity.general_registry — Method

general_registry() -> String

Guess the path of the General registry.

RegistryActivity.get_date — Method

get_date(c)

Get the UTC date of the Git commit c.

RegistryActivity.months_list — Method

months_list(path, start_month, end_month)

Get the range of months when there have been commit to the Git repository at path, between start_month and end_month.

RegistryActivity.registry_activity — Function

registry_activity(path=clone_registry();
                  branch="master",
                  start_month=Date(2018, 2),
                  end_month=Date(year(now(UTC)), month(now(UTC))),
                  filter=p->!(occursin(r"_jll$", last(p)["name"]) || last(p)["name"]=="julia"),
                  )

Get the 3-tuple months, packages, versions of vector of months, number of packages, number of versions for the registry at path (by default a Git clone of the General registry with clone_registry()), checking out the Git branch. The search is restricted between start_month and end_month. You can filter the packages matching the filter condition, which is a function taking as argument an item in the dictionary of a parsed Registry.toml, that is a pair whose key is the UUID of the package and the value is a dictionary with information about the package, usually including the name and the path to the directory where the package is recorded.